Feature/user request

+++ src/main/java/ru/edu/online/services/Impl/PassRequestServiceImpl.java
@@ -226,10 +226,10 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @return список заявок с определенным статусом
      */
     @Override
-    public Optional<PassRequestsResponseDTO> getPassRequestByStatusForUser(String authorId,
-                                                                           String status,
-                                                                           Long page,
-                                                                           Long pageSize) {
+    public Optional<ResponseDTO<PassRequest>> getPassRequestByStatusForUser(String authorId,
+                                                               String status,
+                                                               Long page,
+                                                               Long pageSize) {
         List<PassRequest> requests =
                 passRequestRepository.findAllByAuthorId(authorId);
         log.info("Getting passRequests by status");
@@ -330,11 +330,11 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @return отобранные по критериям заявки
      */
     @Override
-    public Optional<PassRequestsResponseDTO> getPassRequestsForAdmin(RequestsStatusForAdmin status,
-                                                               Long page,
-                                                               Long pageSize,
-                                                               Optional<String> search,
-                                                               Principal principal
+    public Optional<ResponseDTO<PassRequest>> getPassRequestsForAdmin(RequestsStatusForAdmin status,
+                                                         Long page,
+                                                         Long pageSize,
+                                                         String search,
+                                                         Principal principal
     ) {
         UserDTO admin = ScosApiUtils.getUserDetails(devScosApiClient, principal);
         String adminUniversityId = admin
@@ -367,11 +367,11 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @param search поиск
      * @return список заявок в обработке
      */
-    public PassRequestsResponseDTO getPassRequestsForProcessing(
+    public ResponseDTO<PassRequest> getPassRequestsForProcessing(
             String universityId,
             Long page,
             Long pageSize,
-            Optional<String> search) {
+            String search) {
         log.info("collect requests sent for consideration to the target OOVO");
         return aggregatePassRequestsByStatusWithPaginationAndSearchForUniversity(
                 new PassRequestStatus[]{PassRequestStatus.TARGET_ORGANIZATION_REVIEW},
@@ -390,11 +390,11 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @param search поиск
      * @return список заявок в обработке
      */
-    public PassRequestsResponseDTO getPassRequestsInProcessing(
+    public ResponseDTO<PassRequest> getPassRequestsInProcessing(
             String universityId,
             Long page,
             Long pageSize,
-            Optional<String> search) {
+            String search) {
         log.info("collect requests sent in consideration to the target OOVO");
 
         return aggregatePassRequestsByStatusWithPaginationAndSearchForUniversity(
@@ -414,10 +414,10 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @param search поиск
      * @return список обработанных заявок
      */
-    public PassRequestsResponseDTO getExpiredPassRequests(String universityId,
-                                                          Long page,
-                                                          Long pageSize,
-                                                          Optional<String> search) {
+    public ResponseDTO<PassRequest> getExpiredPassRequests(String universityId,
+                                              Long page,
+                                              Long pageSize,
+                                              String search) {
         log.info("collect expired requests sent for to the OOVO");
         return aggregatePassRequestsByStatusWithPaginationAndSearchForUniversity(
                 new PassRequestStatus[]{PassRequestStatus.EXPIRED},
@@ -436,10 +436,10 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @param search поиск
      * @return список обработанных заявок
      */
-    public PassRequestsResponseDTO getProcessedPassRequests(String universityId,
-                                                            Long page,
-                                                            Long pageSize,
-                                                            Optional<String> search) {
+    public ResponseDTO<PassRequest> getProcessedPassRequests(String universityId,
+                                                Long page,
+                                                Long pageSize,
+                                                String search) {
 
         log.info("collect considered requests sent for to the OOVO");
         return aggregatePassRequestsByStatusWithPaginationAndSearchForUniversity(
@@ -492,20 +492,34 @@ public class PassRequestServiceImpl implements IPassRequestService {
         Optional<PassRequest> passRequest = getPassRequest(dto);
 
         if (passRequest.isPresent()) {
-            PassRequestChangeLogEntry entry = new PassRequestChangeLogEntry(
-                    "PassRequestStatus", passRequest.get().getStatus().toString(),
-                    dto.getStatus().toString(), dto.getId()
-            );
-            passRequestChangeLogRepository.save(entry);
-
-            passRequest.get().setStatus(dto.getStatus());
-            passRequestRepository.save(passRequest.get());
-
+            addChangeLogEntryToPassRequest(passRequest.get(), dto.getStatus());
             log.info("pass request with id: {} was updated", dto.getId());
             return getPassRequest(passRequest.get().getId());
         } else
             log.warn("pass request with id: {} not found", dto.getId());
             return Optional.empty();
+    }
+
+    /**
+     * Обновить даты действия заявки
+     * @param passRequestDTO DTO обновленной заявки
+     * @return обновлённая заявка
+     */
+    @Override
+    public Optional<PassRequest> updatePassRequestDates(PassRequestDTO passRequestDTO) {
+        Optional<PassRequest> passRequest = getPassRequest(passRequestDTO.getId());
+         if (passRequest.isPresent()) {
+             if (Optional.ofNullable(passRequestDTO.getStartDate()).isPresent()) {
+                 passRequest.get().setStartDate(passRequestDTO.getStartDate());
+             }
+             if (Optional.ofNullable(passRequestDTO.getEndDate()).isPresent()) {
+                 passRequest.get().setEndDate(passRequestDTO.getEndDate());
+             }
+
+             passRequestRepository.save(passRequest.get());
+             return passRequest;
+         }
+        return Optional.empty();
     }
 
     /**
@@ -597,30 +611,35 @@ public class PassRequestServiceImpl implements IPassRequestService {
     }
 
     /**
-     * Удаление просроченных заявок
-     * @return список удаленных заявок
+     * Поиск и удаление просроченных заявок
+     * Удалению через 7 дней подлежат заявки со статусами:
+     * EXPIRED
+     * CANCELED_BY_CREATOR
+     * REJECTED_BY_TARGET_ORGANIZATION
      */
     @Override
-    public Optional<List<PassRequest>> getExpiredPassRequests() {
-        // Поиск просроченных заявок, изменение их статуса на EXPIRED
+    public void getExpiredPassRequests() {
         checkExpiredPassRequests();
-
-        // Поиск заявок со статусом EXPIRED и CANCELED_BY_CREATOR
         List<PassRequest> expiredList =
                 passRequestRepository.findAllByStatus(PassRequestStatus.EXPIRED);
-        /* Функциональность временно ограничена
+
         expiredList
                 .addAll(
                         passRequestRepository.findAllByStatus(PassRequestStatus.CANCELED_BY_CREATOR)
                 );
+        expiredList
+                .addAll(
+                        passRequestRepository.findAllByStatus(PassRequestStatus.REJECTED_BY_TARGET_ORGANIZATION)
+                );
 
         for (PassRequest request : expiredList) {
-            passRequestRepository.deleteById(request.getId());
+            Optional<PassRequestChangeLogEntry> entry = findInvalidRequests(request);
+            if (entry.isPresent() && entry.get().getDate().isBefore(LocalDate.now().minusDays(7))) {
+                log.info("old pass request with id {} was removed", request.getId());
+                passRequestRepository.deleteById(request.getId());
+            }
         }
-        log.info("expired pass requests has been deleted");
-        */
-        log.info("expired pass requests has been get");
-        return Optional.of(expiredList);
+        deleteOldAcceptedRequests();
     }
 
     /**
@@ -645,6 +664,64 @@ public class PassRequestServiceImpl implements IPassRequestService {
                 status != PassRequestStatus.EXPIRED &&
                 status != PassRequestStatus.CANCELED_BY_CREATOR &&
                 startDate.isBefore(LocalDate.now()));
+    }
+
+    /**
+     * Добавить новую запись в список изменений заявки.
+     * @param passRequest заявка
+     * @param status новый статус
+     */
+    private void addChangeLogEntryToPassRequest(PassRequest passRequest,
+                                                PassRequestStatus status) {
+        PassRequestChangeLogEntry entry = new PassRequestChangeLogEntry(
+                "PassRequestStatus", passRequest.getStatus().toString(),
+                status.toString(), passRequest.getId()
+        );
+        passRequestChangeLogRepository.save(entry);
+
+        passRequest.setStatus(status);
+        passRequestRepository.save(passRequest);
+    }
+
+    /**
+     * Удаление одобренных заявок, срок действия которых истёк более чем на 7 дней
+     */
+    private void deleteOldAcceptedRequests() {
+        List<PassRequest> acceptedRequests =
+                passRequestRepository.findAllByStatus(PassRequestStatus.ACCEPTED);
+
+        for (PassRequest request : acceptedRequests) {
+            if (request.getEndDate().isBefore(LocalDate.now().minusDays(7))) {
+                log.info("delete old accepted pass request with id {}", request.getId());
+                passRequestRepository.deleteById(request.getId());
+            }
+        }
+    }
+
+    /**
+     * Проверка по истории изменений заявки на бесполезность
+     * @param passRequest заявка
+     * @return либо запись из списка изменений о негодности, либо empty()
+     */
+    private Optional<PassRequestChangeLogEntry> findInvalidRequests(PassRequest passRequest) {
+        Optional<PassRequestChangeLogEntry> entry;
+        PassRequestStatus[] statuses = new PassRequestStatus[]{
+                PassRequestStatus.REJECTED_BY_TARGET_ORGANIZATION,
+                PassRequestStatus.CANCELED_BY_CREATOR,
+                PassRequestStatus.EXPIRED
+        };
+
+        for (PassRequestStatus status : statuses) {
+            entry = passRequest.getChangeLog()
+                            .stream()
+                            .filter(log -> log.getNewValue().equals(status.toString()))
+                            .findFirst();
+            if (entry.isPresent()) {
+                return entry;
+            }
+        }
+
+        return Optional.empty();
     }
 
     /**
@@ -812,10 +889,10 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @param pageSize размер страницы     *
      * @return список отобранных заявок по критериям выше
      */
-    private PassRequestsResponseDTO aggregatePassRequestsByStatusWithPaginationForUser(List<PassRequest> requests,
-                                                                                       PassRequestStatus[] statuses,
-                                                                                       Long page,
-                                                                                       Long pageSize) {
+    private ResponseDTO<PassRequest> aggregatePassRequestsByStatusWithPaginationForUser(List<PassRequest> requests,
+                                                                           PassRequestStatus[] statuses,
+                                                                           Long page,
+                                                                           Long pageSize) {
         List<PassRequest> filteredRequest = new LinkedList<>();
         for (PassRequestStatus status : statuses) {
             filteredRequest.addAll(
@@ -824,7 +901,7 @@ public class PassRequestServiceImpl implements IPassRequestService {
                             .collect(Collectors.toList())
             );
         }
-        return new PassRequestsResponseDTO(
+        return new ResponseDTO<>(
                 page,
                 pageSize,
                 filteredRequest.size() % pageSize == 0 ?
@@ -844,12 +921,12 @@ public class PassRequestServiceImpl implements IPassRequestService {
      * @param search поиск (опционально)
      * @return список заявок по входным параметрам
      */
-    private PassRequestsResponseDTO aggregatePassRequestsByStatusWithPaginationAndSearchForUniversity(
+    private ResponseDTO<PassRequest> aggregatePassRequestsByStatusWithPaginationAndSearchForUniversity(
             PassRequestStatus[] statuses,
             String universityId,
             Long page,
             Long pageSize,
-            Optional<String> search) {
+            String search) {
         List<PassRequest> requestList = new LinkedList<>();
         for (PassRequestStatus status : statuses) {
             requestList.addAll(
@@ -860,13 +937,13 @@ public class PassRequestServiceImpl implements IPassRequestService {
             );
         }
 
-        if (search.isPresent()) {
-            requestList = PassRequestUtils.filterRequest(requestList, search.get(), devScosApiClient);
+        if (Optional.ofNullable(search).isPresent()) {
+            requestList = PassRequestUtils.filterRequest(requestList, search, devScosApiClient);
         }
         long requestsCount = requestList.size();
         requestList = paginateRequests(requestList, page, pageSize);
 
-        return new PassRequestsResponseDTO(
+        return new ResponseDTO<>(
                 page,
                 pageSize,
                 requestsCount % pageSize == 0 ? requestsCount / pageSize : requestsCount / pageSize + 1,
